♻refactor: #S10P21C101-backend 평점 스코어 수정
Former-commit-id: ad38b1e287e86b57ac4c1dd9ac5b45612dab309b

+++ backend/src/main/java/com/ssafy/backend/domain/book/dto/response/BookReviewMyResponseDto.java
@@ -8,7 +8,7 @@ public record BookReviewMyResponseDto(
         Long userId,
         String title,
         String coverPath,
-        int score,
+        double score,
         String content
 ) {
 }
